chore(workspace): migrate to Rust edition 2024
Set rust-version = "1.93" and upgrade dependency versions across
workspace. Apply edition 2024 code fixes and normalize
rust-toolchain channel. Remove Cargo.lock.

+++ sqlx-sqlite/src/statement/handle.rs
@@ -335,12 +335,12 @@ impl StatementHandle {
 
     #[inline]
     pub(crate) fn column_int(&self, index: usize) -> i32 {
-        unsafe { sqlite3_column_int(self.0.as_ptr(), check_col_idx!(index)) as i32 }
+        unsafe { sqlite3_column_int(self.0.as_ptr(), check_col_idx!(index)) }
     }
 
     #[inline]
     pub(crate) fn column_int64(&self, index: usize) -> i64 {
-        unsafe { sqlite3_column_int64(self.0.as_ptr(), check_col_idx!(index)) as i64 }
+        unsafe { sqlite3_column_int64(self.0.as_ptr(), check_col_idx!(index)) }
     }
 
     #[inline]
